[feat] StampMissionBasicSetting 엔티티 삭제, 필드 Stamp로 이전 (#176)
